requirements.txt
spotipy==2.23.0
urllib3<2.0.0
Flask==3.0.0
python-dotenv==1.0.0
